Add fix for picking correct Handover connection
A single handover call cannot track both the foreground and
background call state after handover.

Added support to go through all the handover connections and
pick the first match based on the number and if there is no
match check the state.

Added support for tracking the state of call associated with
the connection before handover. This is used to retrieve the
correct connection based on state.

Change-Id: Iec87c09523d459998d0880252a71f98eb8e296a2
CRs-Fixed: 760336